+++ code/src/sixsq/nuvla/server/resources/callback_create_session_github.clj
@@ -36,15 +36,17 @@
               (if matched-user-id
                 (let [{identifier :name} (ex/get-user matched-user-id)
                       cookie-info     (cookies/create-cookie-info matched-user-id
-                                                                 :session-id session-id)
+                                                                  :session-id session-id)
                       cookie          (cookies/create-cookie cookie-info)
                       expires         (ts/rfc822->iso8601 (:expires cookie))
-                      claims-roles    (:claims cookie-info)
+                      claims          (:claims cookie-info)
+                      groups          (:groups cookie-info)
                       updated-session (cond-> (assoc current-session
                                                 :user matched-user-id
                                                 :identifier (or identifier matched-user-id)
                                                 :expiry expires)
-                                              claims-roles (assoc :roles claims-roles))
+                                              claims (assoc :roles claims)
+                                              groups (assoc :groups groups))
                       {:keys [status] :as resp} (sutils/update-session session-id updated-session)]
                   (log/debug "github cookie token claims for" instance ":" cookie-info)
                   (if (not= status 200)