+++ code/src/sixsq/nuvla/server/resources/session/utils.clj
@@ -19,6 +19,7 @@
     (-> (db/retrieve id {})
         (merge body)
         u/update-timestamps
+        (u/set-updated-by request)
         crud/validate
         (db/edit request))
     (catch Exception e